refactor(actions): streamline bash execution messages
